% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dailyret.R
\name{dailyret}
\alias{dailyret}
\title{Daily returns for a company enlisted in the Dhaka Stock Exchange}
\usage{
dailyret(id, summarize = FALSE)
}
\arguments{
\item{id}{The vector containing closing price for any company}
}
\value{
Vector containing the daily returns of the company
}
\description{
This function returns the daily return for any company enlisted in the Dhaka Stock Exchange in any year between 2005 and 2015.
}
\details{
This function takes a vector containing the closing price and returns the daily returns.
}
\author{
Syed M. Fuad
}
\seealso{
\code{length}
}

